Final_code/requirements.txt
# tensorflow>=1.11.0,<2.00.0  # BERT code doesn't work for tensorflow 2.0
# tensorflow-gpu>=1.11.0,<2.00.0  # for gpu
pytorch==1.13.0
torchvision==0.14.0
absl-py>=0.8.1
altair>=3.2.0
attrs>=19.3.0
cycler>=0.10.0
entrypoints>=0.3
importlib-metadata>=0.23
Jinja2>=2.10.3
joblib>=0.14.0
jsonschema>=3.1.1
kiwisolver>=1.1.0
MarkupSafe>=1.1.1
matplotlib>=3.1.1
more-itertools>=7.2.0
numpy>=1.17.2
pandas>=0.25.1
patsy>=0.5.1
praw>=6.4.0
pyparsing>=2.4.2
pyrsistent>=0.15.4
python-dateutil>=2.8.0
pytz>=2019.3
scikit-learn>=0.21.3
scipy>=1.3.1
seaborn>=0.9.0
six>=1.12.0
sklearn>=0.0
statsmodels>=0.10.1
# tf-slim>=1.1.0
toolz>=0.10.0
zipp>=0.6.0
tqdm
